+++ OsmAnd/src/net/osmand/plus/views/POIMapLayer.java
@@ -184,7 +184,7 @@ public class POIMapLayer extends OsmandMapLayer implements ContextMenuLayer.ICon
 
 	public int getRadiusPoi(RotatedTileBox tb) {
 		int r = 0;
-		final float zoom = tb.getZoom() + tb.getZoomScale();
+		final double zoom = tb.getZoom() + tb.getZoomScale();
 		if (zoom < startZoom) {
 			r = 0;
 		} else if (zoom <= 15) {
